requirements.txt
boto3==1.26.80
coverage==7.2.1
cryptography
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-proxy==1.2.2
django-webpack-loader==1.8.1
djangorestframework==3.14.0
djangorestframework-link-header-pagination==0.1.1
drf-yasg==1.21.5
duo-universal==2.0.3
factory-boy==3.2.1
fastapi==0.92.0
gunicorn==20.1.0
icmplib==3.0.3
influxdb-client==1.36.1
openai==0.26.5
Pillow==9.4.0
uvicorn==0.20.0
websockets==10.4
